% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleDeliveryNotice.R
\name{saleDeliveryNotice_apiSave}
\alias{saleDeliveryNotice_apiSave}
\title{发货通知单接口保存功能}
\usage{
saleDeliveryNotice_apiSave(
  api_token = "BEC6002E-C3AE-4947-AD70-212CF2B4218B",
  meta_token = "AD64F20D-6063-4E87-81E8-A24C1751D758",
  FBillNo = "FHTZD22050032"
)
}
\arguments{
\item{api_token}{API口令}

\item{meta_token}{元数据口令}

\item{FBillNo}{单据编号}
}
\value{
返回值
}
\description{
发货通知单接口保存功能
}
\examples{
saleDeliveryNotice_apiSave()
}
