% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrder.R
\name{saleOrder_queryInterId}
\alias{saleOrder_queryInterId}
\title{销售订单接口查询订单及分录行内码}
\usage{
saleOrder_queryInterId(
  api_token = "BEC6002E-C3AE-4947-AD70-212CF2B4218B",
  meta_token = "AD64F20D-6063-4E87-81E8-A24C1751D758",
  FBillNo = "XSDD22050025"
)
}
\arguments{
\item{api_token}{API口令}

\item{meta_token}{元数据口令}

\item{FBillNo}{单据编号}
}
\value{
返回值
}
\description{
销售订单接口查询订单及分录行内码
}
\examples{
saleOrder_apiSave()
}
