% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOutStock.R
\name{saleOutStock_apiSave}
\alias{saleOutStock_apiSave}
\title{销售出库单的保存}
\usage{
saleOutStock_apiSave(
  api_token = "BEC6002E-C3AE-4947-AD70-212CF2B4218B",
  meta_token = "AD64F20D-6063-4E87-81E8-A24C1751D758",
  FBillNo = "XSCKD2022050001"
)
}
\arguments{
\item{api_token}{接口}

\item{meta_token}{元数据}

\item{FBillNo}{单据编码}
}
\value{
返回值
}
\description{
销售出库单的保存
}
\examples{
saleOutStock_apiSave()
}
