% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOutStock.R
\name{saleOutStock_View}
\alias{saleOutStock_View}
\title{销售出库单单据信息查询}
\usage{
saleOutStock_View(
  api_token = "BEC6002E-C3AE-4947-AD70-212CF2B4218B",
  meta_token = "AD64F20D-6063-4E87-81E8-A24C1751D758",
  FBillNo = "XSCKD2022050002"
)
}
\arguments{
\item{api_token}{api口令}

\item{meta_token}{元数据}

\item{FBillNo}{单据编号}
}
\value{
返回值
}
\description{
销售出库单单据信息查询
}
\examples{
saleOutStock_View()
}
